src/main/java/com/mslup/lot/lotcrud/service/FlightService.java
package com.mslup.lot.lotcrud.service;

import com.mslup.lot.lotcrud.model.Flight;
import com.mslup.lot.lotcrud.repository.FlightRepository;
import java.util.List;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;

@Service
@RequiredArgsConstructor
public class FlightService {
    private final FlightRepository flightRepository;

    public Flight addFlight(Flight flight) {
        return flightRepository.save(flight);
    }

    public List<Flight> getAllFlights() {
        return flightRepository.findAll();
    }
}
